+++ core/src/mast/serialization/info.rs
@@ -2,7 +2,9 @@ use miden_crypto::hash::rpo::RpoDigest;
 use winter_utils::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable};
 
 use super::{basic_block_data_decoder::BasicBlockDataDecoder, DataOffset};
-use crate::mast::{MastForest, MastNode, MastNodeId};
+use crate::mast::{
+    BasicBlockNode, CallNode, JoinNode, LoopNode, MastForest, MastNode, MastNodeId, SplitNode,
+};
 
 // MAST NODE INFO
 // ================================================================================================
@@ -30,57 +32,45 @@ impl MastNodeInfo {
         mast_forest: &MastForest,
         basic_block_data_decoder: &BasicBlockDataDecoder,
     ) -> Result<MastNode, DeserializationError> {
-        let mast_node = match self.ty {
+        match self.ty {
             MastNodeType::Block {
                 offset,
                 len: num_operations_and_decorators,
             } => {
                 let (operations, decorators) = basic_block_data_decoder
                     .decode_operations_and_decorators(offset, num_operations_and_decorators)?;
-
-                Ok(MastNode::new_basic_block_with_decorators(operations, decorators))
+                let block = BasicBlockNode::new_unsafe(operations, decorators, self.digest);
+                Ok(MastNode::Block(block))
             },
             MastNodeType::Join { left_child_id, right_child_id } => {
                 let left_child = MastNodeId::from_u32_safe(left_child_id, mast_forest)?;
                 let right_child = MastNodeId::from_u32_safe(right_child_id, mast_forest)?;
-
-                Ok(MastNode::new_join(left_child, right_child, mast_forest)
-                    .expect("invalid node id"))
+                let join = JoinNode::new_unsafe([left_child, right_child], self.digest);
+                Ok(MastNode::Join(join))
             },
             MastNodeType::Split { if_branch_id, else_branch_id } => {
                 let if_branch = MastNodeId::from_u32_safe(if_branch_id, mast_forest)?;
                 let else_branch = MastNodeId::from_u32_safe(else_branch_id, mast_forest)?;
-
-                Ok(MastNode::new_split(if_branch, else_branch, mast_forest)
-                    .expect("invalid node id"))
+                let split = SplitNode::new_unsafe([if_branch, else_branch], self.digest);
+                Ok(MastNode::Split(split))
             },
             MastNodeType::Loop { body_id } => {
                 let body_id = MastNodeId::from_u32_safe(body_id, mast_forest)?;
-
-                Ok(MastNode::new_loop(body_id, mast_forest).expect("invalid node id"))
+                let loop_node = LoopNode::new_unsafe(body_id, self.digest);
+                Ok(MastNode::Loop(loop_node))
             },
             MastNodeType::Call { callee_id } => {
                 let callee_id = MastNodeId::from_u32_safe(callee_id, mast_forest)?;
-
-                Ok(MastNode::new_call(callee_id, mast_forest).expect("invalid node id"))
+                let call = CallNode::new_unsafe(callee_id, self.digest);
+                Ok(MastNode::Call(call))
             },
             MastNodeType::SysCall { callee_id } => {
                 let callee_id = MastNodeId::from_u32_safe(callee_id, mast_forest)?;
-
-                Ok(MastNode::new_syscall(callee_id, mast_forest).expect("invalid node id"))
+                let syscall = CallNode::new_syscall_unsafe(callee_id, self.digest);
+                Ok(MastNode::Call(syscall))
             },
             MastNodeType::Dyn => Ok(MastNode::new_dyn()),
             MastNodeType::External => Ok(MastNode::new_external(self.digest)),
-        }?;
-
-        if mast_node.digest() == self.digest {
-            Ok(mast_node)
-        } else {
-            Err(DeserializationError::InvalidValue(format!(
-                "MastNodeInfo's digest '{}' doesn't match deserialized MastNode's digest '{}'",
-                self.digest,
-                mast_node.digest()
-            )))
         }
     }
 }