+++ processor/src/host/mast_forest_store.rs
@@ -26,7 +26,8 @@ impl MemMastForestStore {
     pub fn insert(&mut self, mast_forest: MastForest) {
         let mast_forest = Arc::new(mast_forest);
 
-        for proc_digest in mast_forest.procedure_digests() {
+        // only register the procedures which are local to this forest
+        for proc_digest in mast_forest.local_procedure_digests() {
             self.mast_forests.insert(proc_digest, mast_forest.clone());
         }
     }